chore(tools): extract rara-background-tasks crate from bash.rs
