+++ src/tools/bash.rs
@@ -1,27 +1,29 @@
 use std::collections::HashMap;
 use std::env;
-use std::io::SeekFrom;
 use std::path::Path;
 use std::path::PathBuf;
 use std::process::Stdio;
 use std::sync::Arc;
-use std::sync::Mutex;
 use std::sync::atomic::{AtomicBool, Ordering};
 use std::time::Duration;
 use std::time::Instant;
 
 use async_trait::async_trait;
+use rara_background_tasks::{
+    BackgroundTaskListTool, BackgroundTaskRecord, BackgroundTaskState, BackgroundTaskStatus,
+    BackgroundTaskStatusTool, BackgroundTaskStopTool, BackgroundTaskStore, BashStreamKind,
+    append_background_output, read_output_tail,
+};
 use rara_tool_macros::tool_spec;
 use rara_tools::tool::{Tool, ToolCallContext, ToolError, ToolOutputStream, ToolProgressEvent};
 use serde::{Deserialize, Serialize};
 use serde_json::{Value, json};
 use tokio::fs;
-use tokio::io::{AsyncRead, AsyncReadExt, AsyncSeekExt, AsyncWriteExt};
+use tokio::io::{AsyncRead, AsyncReadExt, AsyncWriteExt};
 use tokio::process::{Child, Command};
 use tokio::sync::mpsc;
 use tokio::sync::oneshot;
 use tokio::time::{Instant as TokioInstant, MissedTickBehavior};
-use uuid::Uuid;
 
 use crate::sandbox::{SandboxManager, WrappedCommand, sandbox_failure_hint};
 use crate::tool_result::model_preview_bash_output;
@@ -32,87 +34,6 @@ pub struct BashTool {
     pub background_tasks: Arc<BackgroundTaskStore>,
     pub base_env: Arc<HashMap<String, String>>,
     pub sandbox_network_access: Arc<AtomicBool>,
-}
-
-pub struct BackgroundTaskStatusTool {
-    pub background_tasks: Arc<BackgroundTaskStore>,
-}
-
-pub struct BackgroundTaskListTool {
-    pub background_tasks: Arc<BackgroundTaskStore>,
-}
-
-pub struct BackgroundTaskStopTool {
-    pub background_tasks: Arc<BackgroundTaskStore>,
-}
-
-#[derive(Debug, Clone)]
-pub struct BackgroundTaskStore {
-    dir: PathBuf,
-    tasks: Arc<Mutex<HashMap<String, BackgroundTaskRecord>>>,
-    stop_signals: Arc<Mutex<HashMap<String, oneshot::Sender<()>>>>,
-}
-
-#[derive(Debug, Clone, Serialize)]
-pub struct BackgroundTaskRecord {
-    id: String,
-    command: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    program: Option<String>,
-    #[serde(skip_serializing_if = "Vec::is_empty", default)]
-    args: Vec<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    cwd: Option<String>,
-    output_path: PathBuf,
-    status: BackgroundTaskStatus,
-    exit_code: Option<i32>,
-    sandboxed: bool,
-    sandbox_backend: String,
-    network_access: bool,
-}
-
-#[derive(Debug, Clone, Copy, Serialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum BackgroundTaskStatus {
-    Running,
-    Completed,
-    Failed,
-    Killed,
-}
-
-/// Higher-level classifier for background task lifecycle state.
-#[derive(Debug, Clone, Copy, Serialize, PartialEq, Eq)]
-#[serde(rename_all = "snake_case")]
-pub enum BackgroundTaskState {
-    Working,
-    Blocked,
-    Done,
-    Failed,
-}
-
-impl BackgroundTaskStatus {
-    pub fn classify(self) -> BackgroundTaskState {
-        match self {
-            Self::Running | Self::Killed => BackgroundTaskState::Working,
-            Self::Completed => BackgroundTaskState::Done,
-            Self::Failed => BackgroundTaskState::Failed,
-        }
-    }
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-enum BashStreamKind {
-    Stdout,
-    Stderr,
-}
-
-impl BashStreamKind {
-    fn output_stream(self) -> ToolOutputStream {
-        match self {
-            Self::Stdout => ToolOutputStream::Stdout,
-            Self::Stderr => ToolOutputStream::Stderr,
-        }
-    }
 }
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq)]
@@ -431,132 +352,6 @@ fn parse_cd_target(input: &str) -> Option<(String, &str)> {
     Some((path.to_string(), &input[path_end..]))
 }
 
-impl BackgroundTaskStore {
-    pub fn new(dir: PathBuf) -> Result<Self, ToolError> {
-        std::fs::create_dir_all(&dir)?;
-        Ok(Self {
-            dir,
-            tasks: Arc::new(Mutex::new(HashMap::new())),
-            stop_signals: Arc::new(Mutex::new(HashMap::new())),
-        })
-    }
-
-    fn start_record(
-        &self,
-        request: &BashCommandInput,
-        sandboxed: bool,
-        sandbox_backend: String,
-        network_access: bool,
-    ) -> Result<(BackgroundTaskRecord, oneshot::Receiver<()>), ToolError> {
-        let id = format!("bash-{}", Uuid::new_v4());
-        let output_path = self.dir.join(format!("{id}.log"));
-        let record = BackgroundTaskRecord {
-            id: id.clone(),
-            command: request.summary(),
-            program: request
-                .program
-                .as_deref()
-                .filter(|v| !v.trim().is_empty())
-                .map(String::from),
-            args: request.args.clone(),
-            cwd: request
-                .cwd
-                .as_ref()
-                .filter(|d| !d.trim().is_empty())
-                .cloned(),
-            output_path,
-            status: BackgroundTaskStatus::Running,
-            exit_code: None,
-            sandboxed,
-            sandbox_backend,
-            network_access,
-        };
-        let (stop_tx, stop_rx) = oneshot::channel();
-        self.tasks
-            .lock()
-            .expect("background task store lock")
-            .insert(id.clone(), record.clone());
-        self.stop_signals
-            .lock()
-            .expect("background task stop signal lock")
-            .insert(id, stop_tx);
-        Ok((record, stop_rx))
-    }
-
-    fn finish(&self, id: &str, status: BackgroundTaskStatus, exit_code: Option<i32>) {
-        if let Some(record) = self
-            .tasks
-            .lock()
-            .expect("background task store lock")
-            .get_mut(id)
-        {
-            if !matches!(record.status, BackgroundTaskStatus::Killed) {
-                record.status = status;
-            }
-            record.exit_code = exit_code;
-        }
-        self.stop_signals
-            .lock()
-            .expect("background task stop signal lock")
-            .remove(id);
-    }
-
-    fn get(&self, id: &str) -> Option<BackgroundTaskRecord> {
-        self.tasks
-            .lock()
-            .expect("background task store lock")
-            .get(id)
-            .cloned()
-    }
-
-    fn list(&self) -> Vec<BackgroundTaskRecord> {
-        let mut records = self
-            .tasks
-            .lock()
-            .expect("background task store lock")
-            .values()
-            .cloned()
-            .collect::<Vec<_>>();
-        records.sort_by(|left, right| left.id.cmp(&right.id));
-        records
-    }
-
-    fn stop(&self, id: &str) -> Result<BackgroundTaskRecord, ToolError> {
-        let mut tasks = self.tasks.lock().expect("background task store lock");
-        let record = tasks
-            .get_mut(id)
-            .ok_or_else(|| ToolError::InvalidInput(format!("unknown task id: {id}")))?;
-        if !matches!(record.status, BackgroundTaskStatus::Running) {
-            return Ok(record.clone());
-        }
-        record.status = BackgroundTaskStatus::Killed;
-        let stopped = record.clone();
-        drop(tasks);
-
-        if let Some(stop) = self
-            .stop_signals
-            .lock()
-            .expect("background task stop signal lock")
-            .remove(id)
-        {
-            let _ = stop.send(());
-        }
-        Ok(stopped)
-    }
-
-    fn stop_all(&self) -> Vec<BackgroundTaskRecord> {
-        let ids = self
-            .list()
-            .into_iter()
-            .filter(|record| matches!(record.status, BackgroundTaskStatus::Running))
-            .map(|record| record.id)
-            .collect::<Vec<_>>();
-        ids.into_iter()
-            .filter_map(|id| self.stop(&id).ok())
-            .collect()
-    }
-}
-
 fn sandbox_command_env(
     sandbox_home: &Path,
     base_env: &HashMap<String, String>,
@@ -795,7 +590,18 @@ impl Tool for BashTool {
         let sandbox_perm = request.sandbox_permissions;
         if request.run_in_background {
             let (record, stop_rx) = self.background_tasks.start_record(
-                &request,
+                request.summary(),
+                request
+                    .program
+                    .as_deref()
+                    .filter(|v| !v.trim().is_empty())
+                    .map(String::from),
+                request.args.clone(),
+                request
+                    .cwd
+                    .as_ref()
+                    .filter(|d| !d.trim().is_empty())
+                    .cloned(),
                 wrapped.sandboxed,
                 wrapped.sandbox_backend.clone(),
                 wrapped.network_access,
@@ -974,101 +780,6 @@ impl Tool for BashTool {
     }
 }
 
-#[tool_spec(
-    name = "background_task_list",
-    description = "List background bash tasks started with bash run_in_background. Use this before starting duplicate long-running work when task state is unclear.",
-    input_schema = {
-        "type": "object",
-        "properties": {}
-    }
-)]
-#[async_trait]
-impl Tool for BackgroundTaskListTool {
-    async fn call(&self, _input: Value) -> Result<Value, ToolError> {
-        Ok(json!({
-            "tasks": self.background_tasks.list(),
-        }))
-    }
-}
-
-#[tool_spec(
-    name = "background_task_status",
-    description = "Inspect a background bash task started with bash run_in_background and read the tail of its output.",
-    input_schema = {
-        "type": "object",
-        "properties": {
-            "task_id": {
-                "type": "string",
-                "description": "Background task id returned by bash run_in_background."
-            },
-            "tail_bytes": {
-                "type": "integer",
-                "minimum": 1,
-                "default": 12000,
-                "description": "Maximum number of output bytes to return from the end of the task log."
-            }
-        },
-        "required": ["task_id"]
-    }
-)]
-#[async_trait]
-impl Tool for BackgroundTaskStatusTool {
-    async fn call(&self, input: Value) -> Result<Value, ToolError> {
-        let task_id = input["task_id"]
-            .as_str()
-            .ok_or_else(|| ToolError::InvalidInput("task_id".into()))?;
-        let tail_bytes = input
-            .get("tail_bytes")
-            .and_then(Value::as_u64)
-            .unwrap_or(12_000)
-            .min(1_000_000) as usize;
-        let record = self
-            .background_tasks
-            .get(task_id)
-            .ok_or_else(|| ToolError::InvalidInput(format!("unknown task id: {task_id}")))?;
-        let output = read_output_tail(&record.output_path, tail_bytes).await?;
-
-        Ok(json!({
-            "task_id": record.id,
-            "command": record.command,
-            "program": record.program,
-            "args": record.args,
-            "cwd": record.cwd,
-            "status": record.status,
-            "exit_code": record.exit_code,
-            "output_path": record.output_path,
-            "output": output,
-            "sandboxed": record.sandboxed,
-            "sandbox_backend": record.sandbox_backend,
-            "network_access": record.network_access,
-        }))
-    }
-}
-
-#[tool_spec(
-    name = "background_task_stop",
-    description = "Stop one background bash task, or all running background bash tasks when task_id is omitted.",
-    input_schema = {
-        "type": "object",
-        "properties": {
-            "task_id": {
-                "type": "string",
-                "description": "Background task id returned by bash run_in_background. Omit to stop all running background bash tasks."
-            }
-        }
-    }
-)]
-#[async_trait]
-impl Tool for BackgroundTaskStopTool {
-    async fn call(&self, input: Value) -> Result<Value, ToolError> {
-        if let Some(task_id) = input.get("task_id").and_then(Value::as_str) {
-            let task = self.background_tasks.stop(task_id)?;
-            return Ok(json!({ "stopped": [task] }));
-        }
-        Ok(json!({ "stopped": self.background_tasks.stop_all() }))
-    }
-}
-
 fn spawn_background_bash_task(
     mut child: Child,
     wrapped: WrappedCommand,
@@ -1189,40 +900,6 @@ async fn run_background_bash_task(
         let _ = fs::remove_file(path).await;
     }
     Ok(status.code())
-}
-
-async fn append_background_output(
-    path: &Path,
-    stream: BashStreamKind,
-    chunk: &str,
-) -> Result<(), ToolError> {
-    let mut file = fs::OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(path)
-        .await?;
-    match stream {
-        BashStreamKind::Stdout => file.write_all(chunk.as_bytes()).await?,
-        BashStreamKind::Stderr => {
-            file.write_all(b"[stderr] ").await?;
-            file.write_all(chunk.as_bytes()).await?;
-        }
-    }
-    Ok(())
-}
-
-async fn read_output_tail(path: &Path, max_bytes: usize) -> Result<String, ToolError> {
-    let mut file = match fs::File::open(path).await {
-        Ok(file) => file,
-        Err(err) if err.kind() == std::io::ErrorKind::NotFound => return Ok(String::new()),
-        Err(err) => return Err(err.into()),
-    };
-    let file_len = file.metadata().await?.len();
-    let start = file_len.saturating_sub(max_bytes as u64);
-    file.seek(SeekFrom::Start(start)).await?;
-    let mut bytes = Vec::with_capacity(max_bytes.min(file_len as usize));
-    file.read_to_end(&mut bytes).await?;
-    Ok(String::from_utf8_lossy(&bytes).into_owned())
 }
 
 async fn read_stream_chunks<R>(
@@ -1355,15 +1032,18 @@ mod tests {
     };
     use std::time::Duration;
 
+    use rara_background_tasks::{
+        BackgroundTaskListTool, BackgroundTaskStatus, BackgroundTaskStatusTool,
+        BackgroundTaskStopTool, BackgroundTaskStore, BashStreamKind, read_output_tail,
+    };
     use rara_tools::tool::{Tool, ToolCallContext, ToolOutputStream, ToolProgressEvent};
     use serde_json::{Value, json};
     use tempfile::tempdir;
 
     use super::{
-        BackgroundTaskListTool, BackgroundTaskStatus, BackgroundTaskStatusTool,
-        BackgroundTaskStopTool, BackgroundTaskStore, BashCommandInput, BashSandboxPermissions,
-        BashStreamKind, BashTool, append_aggregated_bash_output, command_env_for_wrapped,
-        read_output_tail, sandbox_command_env, sandbox_output_hint, unsandboxed_execution_warning,
+        BashCommandInput, BashSandboxPermissions, BashTool, append_aggregated_bash_output,
+        command_env_for_wrapped, sandbox_command_env, sandbox_output_hint,
+        unsandboxed_execution_warning,
     };
     use crate::sandbox::{SandboxManager, WrappedCommand};
     use crate::tool_result::model_preview_bash_output;
@@ -2276,30 +1956,6 @@ mod tests {
                 .and_then(Value::as_bool),
             Some(wrapped.network_access)
         );
-    }
-
-    #[tokio::test]
-    async fn read_output_tail_returns_only_requested_suffix() {
-        let temp = tempdir().expect("tempdir");
-        let path = temp.path().join("task.log");
-        tokio::fs::write(&path, b"0123456789tail")
-            .await
-            .expect("write log");
-
-        let output = read_output_tail(&path, 4).await.expect("tail");
-
-        assert_eq!(output, "tail");
-    }
-
-    #[tokio::test]
-    async fn read_output_tail_missing_file_is_empty() {
-        let temp = tempdir().expect("tempdir");
-
-        let output = read_output_tail(&temp.path().join("missing.log"), 4)
-            .await
-            .expect("missing tail");
-
-        assert_eq!(output, "");
     }
 
     fn binary_exists(program: &str) -> bool {
